Task notification

+++ app/src/main/java/com/example/notes/data/repositories/notes/NoteRepositoryImpl.kt
@@ -8,11 +8,11 @@ class NoteRepositoryImpl(private val noteDao: NoteDao) : NoteRepository {
 
     override suspend fun getAllNotes(): Flow<List<Note>> = noteDao.getAllNotes()
 
-    override suspend fun getNoteById(noteId: Long): Note = noteDao.getNoteById(noteId)
+    override suspend fun getNoteById(noteId: Int): Note = noteDao.getNoteById(noteId)
 
     override suspend fun addNote(note: Note) = noteDao.insertNote(note)
 
     override suspend fun updateNote(note: Note) = noteDao.updateNote(note)
 
-    override suspend fun deleteNotes(notes: List<Long>) = noteDao.deleteNotes(notes)
+    override suspend fun deleteNotes(notes: List<Int>) = noteDao.deleteNotes(notes)
 }
